Restructuring components, updating requirements.txt
