+++ client/app/app/components/sign-up-button.tsx
@@ -1,0 +1,9 @@
+import { SignInButton } from "@clerk/react-router"
+
+export default function CustomSignUpButton() {
+  return (
+    <SignInButton forceRedirectUrl='/signup'>
+      <button className="bg-white">Custom sign in button</button>
+    </SignInButton>
+  )
+}
