+++ client/app/app/components/trip_card.tsx
@@ -1,0 +1,9 @@
+export default function TripCard() {
+    return (
+        <div className="card card-border bg-base-100 size-50 ">
+            <div className="card-body">
+                <h2 className="card-title">Trip 1</h2>
+            </div>
+        </div>
+    )
+}
