--- client/app/app/dashboard/dashboard.tsx
@@ -1,42 +1,0 @@
-import CreateNewButton from "./create_new"
-
-export default function DashboardComponent() {
-    return (
-        <main className="flex items-center justify-center pt-16 pb-4 md:px-40 px-20">
-            <div className="layout-content-container flex flex-col max-w-[960px] flex-1">
-                <div
-                    className="flex min-h-[480px] flex-col gap-6 md:gap-8 bg-cover bg-center bg-no-repeat items-start justify-end px-4 pb-10 md:px-10 rounded-xl"
-                    style={{
-                        backgroundImage: `linear-gradient(rgba(0, 0, 0, 0.1), rgba(0, 0, 0, 0.4)), url("/bg_dash.svg")`
-                    }}
-                    >
-                    <h1>Plan your next travel</h1>
-                </div>
-                <a href="https://www.freepik.com/" className="text-xs text-right">Beach Vectors by <span className="text-blue-500">Freepik</span></a>
-                <div className="grid grid-autofit gap-7 pt-10">
-                    <CreateNewButton />
-                    <div className="card card-border bg-base-100 size-50 ">
-                        <div className="card-body">
-                            <h2 className="card-title">Trip 1</h2>
-                        </div>
-                    </div>
-                    <div className="card card-border bg-base-100 size-50 ">
-                        <div className="card-body">
-                            <h2 className="card-title">Trip 2</h2>
-                        </div>
-                    </div>
-                    <div className="card card-border bg-base-100 size-50 ">
-                        <div className="card-body">
-                            <h2 className="card-title">Trip 3</h2>
-                        </div>
-                    </div>
-                    <div className="card card-border bg-base-100 size-50 ">
-                        <div className="card-body">
-                            <h2 className="card-title">Trip 4</h2>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </main>
-    )
-}
