--- client/app/app/dashboard/trip_card.tsx
@@ -1,5 +1,0 @@
-export default function TripCard() {
-    return (
-        <h1>Hello</h1>
-    )
-}
