+++ server/app/requirements.txt
@@ -1,12 +1,17 @@
 annotated-types==0.7.0
 blinker==1.9.0
+certifi==2025.1.31
+charset-normalizer==3.4.1
 click==8.1.8
 Flask==3.1.0
+idna==3.10
 itsdangerous==2.2.0
 Jinja2==3.1.6
 MarkupSafe==3.0.2
 pydantic==2.10.6
 pydantic_core==2.27.2
 python-dotenv==1.0.1
+requests==2.32.3
 typing_extensions==4.12.2
+urllib3==2.3.0
 Werkzeug==3.1.3
